fix(android): preserve native root mount paths

+++ ish-qt-qml/upstream/ish-ios/fs/generic.c
@@ -12,6 +12,8 @@
 
 struct mount *find_mount_and_trim_path(char *path) {
     struct mount *mount = mount_find(path);
+    if (IS_ERR(mount))
+        return mount;
     char *dst = path;
     const char *src = path + strlen(mount->point);
     while (*src != '\0')
@@ -42,6 +44,8 @@ struct fd *generic_openat(struct fd *at, const char *path_raw, int flags, int mo
     if (err < 0)
         return ERR_PTR(err);
     struct mount *mount = find_mount_and_trim_path(path);
+    if (IS_ERR(mount))
+        return ERR_PTR(PTR_ERR(mount));
     struct fd *fd = mount->fs->open(mount, path, flags, mode);
     if (IS_ERR(fd)) {
         // if an error happens after this point, fd_close will release the
@@ -122,6 +126,8 @@ int generic_accessat(struct fd *dirfd, const char *path_raw, int mode) {
         return err;
 
     struct mount *mount = find_mount_and_trim_path(path);
+    if (IS_ERR(mount))
+        return PTR_ERR(mount);
     struct statbuf stat = {};
     err = mount->fs->stat(mount, path, &stat);
     mount_release(mount);
@@ -140,7 +146,13 @@ int generic_linkat(struct fd *src_at, const char *src_raw, struct fd *dst_at, co
     if (err < 0)
         return err;
     struct mount *mount = find_mount_and_trim_path(src);
+    if (IS_ERR(mount))
+        return PTR_ERR(mount);
     struct mount *dst_mount = find_mount_and_trim_path(dst);
+    if (IS_ERR(dst_mount)) {
+        mount_release(mount);
+        return PTR_ERR(dst_mount);
+    }
     if (mount != dst_mount)
         err = _EXDEV;
     else if (mount->fs->link == NULL)
@@ -158,6 +170,8 @@ int generic_unlinkat(struct fd *at, const char *path_raw) {
     if (err < 0)
         return err;
     struct mount *mount = find_mount_and_trim_path(path);
+    if (IS_ERR(mount))
+        return PTR_ERR(mount);
     err = _EPERM;
     if (mount->fs->unlink)
         err = mount->fs->unlink(mount, path);
@@ -177,7 +191,13 @@ int generic_renameat(struct fd *src_at, const char *src_raw, struct fd *dst_at, 
     if (contains_mount_point(src))
         return _EBUSY;
     struct mount *mount = find_mount_and_trim_path(src);
+    if (IS_ERR(mount))
+        return PTR_ERR(mount);
     struct mount *dst_mount = find_mount_and_trim_path(dst);
+    if (IS_ERR(dst_mount)) {
+        mount_release(mount);
+        return PTR_ERR(dst_mount);
+    }
     if (mount != dst_mount)
         err = _EXDEV;
     else if (mount->fs->rename == NULL)
@@ -195,6 +215,8 @@ int generic_symlinkat(const char *target, struct fd *at, const char *link_raw) {
     if (err < 0)
         return err;
     struct mount *mount = find_mount_and_trim_path(link);
+    if (IS_ERR(mount))
+        return PTR_ERR(mount);
     err = _EPERM;
     if (mount->fs->symlink)
         err = mount->fs->symlink(mount, target, link);
@@ -213,6 +235,8 @@ int generic_mknodat(struct fd *at, const char *path_raw, mode_t_ mode, dev_t_ de
     if (err < 0)
         return err;
     struct mount *mount = find_mount_and_trim_path(path);
+    if (IS_ERR(mount))
+        return PTR_ERR(mount);
     err = _EPERM;
     if (mount->fs->mknod)
         err = mount->fs->mknod(mount, path, mode, dev);
@@ -226,6 +250,8 @@ int generic_setattrat(struct fd *at, const char *path_raw, struct attr attr, boo
     if (err < 0)
         return err;
     struct mount *mount = find_mount_and_trim_path(path);
+    if (IS_ERR(mount))
+        return PTR_ERR(mount);
     err = _EPERM;
     if (mount->fs->setattr)
         err = mount->fs->setattr(mount, path, attr);
@@ -239,6 +265,8 @@ int generic_utime(struct fd *at, const char *path_raw, struct timespec atime, st
     if (err < 0)
         return err;
     struct mount *mount = find_mount_and_trim_path(path);
+    if (IS_ERR(mount))
+        return PTR_ERR(mount);
     err = _EPERM;
     if (mount->fs->utime)
         err = mount->fs->utime(mount, path, atime, mtime);
@@ -252,6 +280,8 @@ ssize_t generic_readlinkat(struct fd *at, const char *path_raw, char *buf, size_
     if (err < 0)
         return err;
     struct mount *mount = find_mount_and_trim_path(path);
+    if (IS_ERR(mount))
+        return PTR_ERR(mount);
     err = _EINVAL;
     if (mount->fs->readlink)
         err = mount->fs->readlink(mount, path, buf, bufsize);
@@ -265,6 +295,8 @@ int generic_mkdirat(struct fd *at, const char *path_raw, mode_t_ mode) {
     if (err < 0)
         return err;
     struct mount *mount = find_mount_and_trim_path(path);
+    if (IS_ERR(mount))
+        return PTR_ERR(mount);
     err = _EPERM;
     if (mount->fs->mkdir)
         err = mount->fs->mkdir(mount, path, mode);
@@ -280,6 +312,8 @@ int generic_rmdirat(struct fd *at, const char *path_raw) {
     if (contains_mount_point(path))
         return _EBUSY;
     struct mount *mount = find_mount_and_trim_path(path);
+    if (IS_ERR(mount))
+        return PTR_ERR(mount);
     err = _EPERM;
     if (mount->fs->rmdir)
         err = mount->fs->rmdir(mount, path);
